BIG step after long session with Sebastian
the two dictionaries I made are kept mainly how they are ad uses to
create rxncon objekt.  In order to do so, the Information put of
reactionID need to be completed by 'Reaction[FULL]' information.  For
this task is started to create a build_full() function (line 371~~)
This is not working correctly yet (compare output to input file)

If this works I need to validate this output by comparing it to the
target column from contID.
For this the function with working tile magic() is there.

If all of that is done, I can create the rxncon object as allready
written but commented out and print it, for example into a text file